refactoring: 向着工具库的方向前进（可以顺利跑起来了）
